Add travis file to automate deployment
The idea here is that travis will run remotely for us what we do run in docker locally

The Docker login username and password will be set as environment variables in the Travis CI account.

The login line is set up in a way that we dont print the password! If you have any variables being passed in, you should be concious whether these variales are sensitive strings that we dont want to be exposed inside of our logs